+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Recording input/output streams.
+
 ## [0.5.0] - 2021-01-18
 ### Added
 - Error indication